refactor: move getDelayInSeconds
