+++ backend-nest/src/utils/fetch.ts
@@ -1,5 +1,3 @@
-import { Delay } from "src/types/types";
-
 export const GOLEMIO_ENDPOINT = new URL(
     "/v2/pid/departureboards",
     "https://api.golemio.cz",
@@ -10,21 +8,4 @@ export const getGolemioHeaders = () => {
         "Content-Type": "application/json",
         "X-Access-Token": process.env.GOLEMIO_API_KEY,
     });
-};
-
-export const getDelayInSeconds = (delay?: Delay | null): number => {
-    if (!delay) {
-        return 0;
-    }
-
-    let seconds = 0;
-
-    if (typeof delay?.seconds === "number") {
-        seconds += delay.seconds;
-    }
-    if (typeof delay?.minutes === "number") {
-        seconds += delay.minutes * 60;
-    }
-
-    return seconds;
 };
